+++ packages/gateway/src/gateway-language-model-settings.ts
@@ -58,6 +58,7 @@ export type GatewayModelId =
   | 'google/gemini-3-pro-image'
   | 'google/gemini-3-pro-preview'
   | 'google/gemini-3.1-flash-image-preview'
+  | 'google/gemini-3.1-flash-lite-preview'
   | 'google/gemini-3.1-pro-preview'
   | 'inception/mercury-coder-small'
   | 'kwaipilot/kat-coder-pro-v1'